+++ src/snitch/mod.rs
@@ -1,7 +1,6 @@
 //  SPDX-FileCopyrightText: Copyright 2023 James M. Putnam (putnamjm.design@gmail.com)
 //  SPDX-License-Identifier: MIT
-// mod button_builder;
-mod status_bar;
+mod host;
+pub mod snitch_ui;
 mod style;
-pub mod tab_ui;
-mod tabs;
+mod tty;
