+++ src/snitch/status_bar.rs
@@ -3,13 +3,13 @@
 #![allow(clippy::collapsible_match)]
 #![allow(unused_imports)]
 use {
-    super::tab_ui::Message,
+    super::snitch_ui::Message,
     crate::Environment,
     iced::{
         executor,
         keyboard::Event::CharacterReceived,
         subscription, theme,
-        widget::{button, column, container, row, rule, text, Column, Space, Text},
+        widget::{button, column, container, horizontal_rule, row, text, Column, Space, Text},
         window, Alignment, Application, Color, Command, Element, Event, Length, Subscription,
         Theme,
     },
@@ -22,7 +22,7 @@ pub struct StatusBar {
 }
 
 impl StatusBar {
-    pub fn new(env: &Environment, _style: String) -> Self {
+    pub fn new(env: &Environment) -> Self {
         let host_path = match &env.hosts_path {
             Some(path) => path.to_str().unwrap().to_string(),
             None => "~/.rsnitch-rs/hosts.json".to_string(),
@@ -31,16 +31,12 @@ impl StatusBar {
         StatusBar { host_path }
     }
 
-    pub fn view(&self) -> Element<Message> {
-        let status_bar = row![
-            text("rsnitch-rs: v0.0.2".to_string()).size(20),
-            text(self.host_path.clone()).size(20),
-        ]
-        .spacing(20);
+    pub fn content(&self, filter: String) -> Element<Message> {
+        let status_bar = text(format!("{}    {}", self.host_path.clone(), filter)).size(20);
 
         let content = Column::new()
             .align_items(Alignment::Start)
-            .spacing(20)
+            .spacing(10)
             .push(status_bar);
 
         container(content)
