--- src/snitch/tabs/controls/mod.rs
@@ -1,3 +1,0 @@
-//  SPDX-FileCopyrightText: Copyright 2023 James M. Putnam (putnamjm.design@gmail.com)
-//  SPDX-License-Identifier: MIT
-pub mod controls_box;
