--- src/snitch/tabs/hosts/mod.rs
@@ -1,7 +1,0 @@
-//  SPDX-FileCopyrightText: Copyright 2023 James M. Putnam (putnamjm.design@gmail.com)
-//  SPDX-License-Identifier: MIT
-pub mod group_box;
-pub mod host;
-pub mod host_box;
-pub mod info_box;
-pub mod tty;
