--- src/snitch/tabs/mod.rs
@@ -1,5 +1,0 @@
-// pub mod controls;
-// pub mod controls_tab;
-pub mod hosts;
-pub mod hosts_tab;
-pub mod settings_tab;
